remove cpu limit in helm chart (#626)
* update reset-minikube
* make clippy happy for new rust
* remove cpu limit in helm chart

+++ src/sc-schema/src/objects/create.rs
@@ -63,7 +63,7 @@ mod create {
 
     impl Encoder for AllCreatableSpec {
         fn write_size(&self, version: Version) -> usize {
-            let type_size = (0 as u8).write_size(version);
+            let type_size = (0u8).write_size(version);
 
             type_size
                 + match self {
